backend/packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mykhailo Razbeiko", email = "mykhailo.razb@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "common",
    "python-multipart>=0.0.22",
    "pyjwt>=2.12.1",
    "pwdlib[argon2]>=0.3.0",
]

[tool.uv.sources]
common = { workspace = true }

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
